Store daily temps with tenth-degree precision
Migrate daily/current/snapshot temperature persistence from integer-backed
storage to float-backed storage so source precision is preserved end-to-end.

Data model and schema updates:
- Change WeatherEntity high/low/current temps to Float?
- Change ForecastSnapshotEntity high/low temps to Float?
- Bump WeatherDatabase to v18 and add MIGRATION_17_18 converting
  weather_data and forecast_snapshots temperature columns from INTEGER to REAL
- Export new Room schema file: app/schemas/.../18.json

Ingestion and repository updates:
- Parse Open-Meteo and WeatherApi daily/current temperatures as Float
  instead of round-to-int values
- Update WeatherRepository temperature pipelines and map types to carry Float
  values without truncation
- Preserve climate normals behavior by converting integer normals to float
  at WeatherEntity mapping boundaries

UI/stat touchpoint updates:
- Update daily text formatting to render integer labels when exact, and
  one decimal when meaningful
- Adjust ForecastHistoryActivity and AccuracyCalculator interfaces for new
  float-backed temperatures while preserving existing chart/stat behavior

Test updates and verification:
- Update affected unit and instrumented tests for float temperature types
  and precision expectations
- Verified with :app:compileDebugKotlin, :app:testDebugUnitTest,
  and emulator instrumented tests via ./scripts/emulator-tests.sh
  (97/97 passed).

+++ app/src/main/java/com/weatherwidget/data/local/ForecastSnapshotEntity.kt
@@ -13,8 +13,8 @@ data class ForecastSnapshotEntity(
     val forecastDate: String, // When forecast was made (e.g., "2024-01-14" for 1-day-ahead)
     val locationLat: Double,
     val locationLon: Double,
-    val highTemp: Int?,
-    val lowTemp: Int?,
+    val highTemp: Float?,
+    val lowTemp: Float?,
     val condition: String,
     val source: String, // Database storage: "NWS", "OPEN_METEO", or "Generic". Use WeatherSource.fromId() to convert.
     val fetchedAt: Long = System.currentTimeMillis(),
